+++ kernel/inc/cap_fs.h
@@ -3,7 +3,7 @@
 #include "error.h"
 
 err_t path_read(cap_t path, char *buf, size_t n);
-err_t path_derive(cte_t src, cte_t dst, const char *path, uint16_t space, path_flags_t flags);
+err_t path_derive(cte_t src, cte_t dst, const char *path, uint32_t space, path_flags_t flags);
 err_t read_file(cap_t path, uint32_t offset, uint8_t *buf, uint32_t buf_size, uint32_t *bytes_read);
 err_t write_file(cap_t path, uint32_t offset, uint8_t *buf, uint32_t buf_size,
 		 uint32_t *bytes_written);
