+++ kernel/src/cap_util.c
@@ -69,7 +69,7 @@ cap_t cap_mk_socket(chan_t chan, ipc_mode_t mode, ipc_perm_t perm, uint32_t tag)
 	return cap;
 }
 
-cap_t cap_mk_path(uint32_t tag, uint16_t space, path_flags_t flags)
+cap_t cap_mk_path(uint32_t tag, uint32_t space, path_flags_t flags)
 {
 	cap_t cap;
 	cap.path.type = CAPTY_PATH;
